% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration-setting-generation.R
\name{configuration_setting_generation}
\alias{configuration_setting_generation}
\title{configuration setting generations}
\usage{
configuration_setting_generation(
  featureNum = 50,
  DataNum = c(100, 100, 100, 100)
)
}
\arguments{
\item{featureNum}{The dimension of the features.}

\item{DataNum}{The number of samples of each dataset.}
}
\description{
This function allows you to set configuration settings manually.
}
\examples{
configuration_setting_generation()
}
\keyword{configuration}
