% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes.R
\name{constr_bugs_model}
\alias{constr_bugs_model}
\title{Contruct Bugs Model}
\usage{
constr_bugs_model(N, n_manifest, n_beta, nclasses, regression, response)
}
\arguments{
\item{regression}{Expression which contains code for the response distribution,
e.g. expr(stuff)}

\item{x}{model matrix}
}
\value{
R function which contains Bugs model
}
\description{
Construct bugs latent class model in the form of a function for use in the code
function bayes_lca
}
