% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{data_sim}
\alias{data_sim}
\title{Simulate data}
\usage{
data_sim(n = 100, k = 12)
}
\arguments{
\item{n}{is number of observations}

\item{k}{number of manifest variables}
}
\description{
Generate a simulated data set
}
