% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{express}
\alias{express}
\title{Small simulated data set}
\format{An object of class \code{data.frame} with 150 rows and 8 columns.}
\usage{
express
}
\description{
Simulated data set with continuous regression outcome. The data set contains
150 observations of 8 variables, which include 5 manifest variables,
and two regressors.
}
\details{
\itemize{
\item{y }{Discrete regression outcome of interest}
\item{Z1 }{Categorical manifest variable 1}
\item{Z2 }{Categorical manifest variable 2}
\item{Z3 }{Categorical manifest variable 3}
\item{Z4 }{Categorical manifest variable 4}
\item{Z5 }{Categorical manifest variable 5}
\item{x1 }{Continuous predictor variable}
\item{x2 }{Continuous predictor variable}
}
}
\keyword{datasets}
