% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcra.R
\name{get_bugs_model}
\alias{get_bugs_model}
\title{Get the Bugs model}
\usage{
get_bugs_model(fit)
}
\arguments{
\item{fit}{an lcra fit object}
}
\value{
R function which contains Bugs model
}
\description{
Sometimes the user may want more flexibility in the model fit than
our program provides. In this case, the user can fit a close model and
use this function to retrieve the model as an R function.
}
