% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCA.R
\name{LCA}
\alias{LCA}
\title{Linked Component Analysis(LCA) function.}
\usage{
LCA(dataset, group, factor_num, T = 100)
}
\arguments{
\item{dataset}{A list containing all the datasets}

\item{group}{A list containing the linked groups of datasets}

\item{factor_num}{A vector contains the number of factors each group should have}

\item{T}{The maximum number of iterations allowed to have in the algorithm}
}
\description{
This function allows you to implement the main algorithm LCA
}
\examples{
configuration_setting = configuration_setting_generation(50, c(100, 100, 100, 100))
dataset = simulated_data_generation(configuration_setting)
group = list(c(1,2,3,4), c(1,2), c(1,4), c(2,3), c(3,4), c(1), c(2), c(3), c(4))
factor_num = c(2, 2, 2, 2, 2, 2, 2, 2, 2)
LCA(dataset, group, factor_num)
}
\keyword{LCA}
