% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes.R
\name{lcra}
\alias{lcra}
\title{Bayesian Latent Class Analysis}
\usage{
lcra(
  formula,
  family,
  data,
  nclasses,
  manifest,
  inits,
  return.bugs,
  dir,
  n.chains,
  n.iter,
  parameters.to.save,
  ...
)
}
\arguments{
\item{formula}{If formula = NULL, LCA without regression model is fitted. If
a regression model is to be fitted, specify a formula using R standard syntax,
e.g., Y ~ age + sex + trt. Do not include manifest variables in the regression
model specification. These will be appended internally as latent classes.}

\item{family}{a description of the error distribution to 
be used in the model. Currently the options are c("gaussian") with identity 
link and c("binomial") which uses a logit link.}

\item{manifest}{character vector containing the names of each manifest variable,
e.g., manifest = c("Z1", "med_3", "X5")}

\item{inits}{list of initial values for WinBUGS. Defaults will be set if nothing
is specified.}

\item{return.bugs}{= FALSE. If TRUE, return the bugs model as a character string.
This can be further edited to suit the researchers needs, e.g., by adding hyper-prior
distributions.}

\item{dir}{Specify full path to the directory where you want
to store the WinBUGS output files and BUGS model file.}

\item{...}{}
}
\value{
a named list of draws.
}
\description{
Perform Bayesian LCA
}
\details{
Given a set of categorical predictors, draw posterior distribution of 
probability of class membership for each observation.
}
