% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paper_sim}
\alias{paper_sim}
\title{Simulated data set}
\format{An object of class \code{data.frame} with 100 rows and 13 columns.}
\usage{
paper_sim
}
\description{
A simulated data set from the paper
}
\details{
List columns:
}
\keyword{datasets}
