% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcra.R
\name{pi}
\alias{pi}
\title{Compute conditional probability of the manifest outcome}
\usage{
pi(lcra.fit)
}
\arguments{
\item{lcra.fit}{the lcra fit object. See ?lcra for instructions.}
}
\value{
a list which contains mean.pi - the posterior mean probability of the manifest outcome,
and pi_draws - the array containing draws for the probability
of the outcome given the implied latent class membership.
}
\description{
This function takes an lcra fit object and computes the probability
of the outcome given the implied latent class membership.
}
