% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcra.R
\name{plot.lcra}
\alias{plot.lcra}
\title{Plotting an lcra object}
\usage{
\method{plot}{lcra}(x, ...)
}
\arguments{
\item{x}{an object of class 'lcra', see lcra() for details}

\item{...}{further arguments to plot}
}
\description{
Plot the lcra model output using plot.bugs.
}
