% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcra.R
\name{print.lcra}
\alias{print.lcra}
\title{Printing an lcra object}
\usage{
\method{print}{lcra}(x, ...)
}
\arguments{
\item{x}{an object of class 'lcra', see lcra() for details}

\item{...}{further arguments to print}
}
\description{
Print the lcra model output using print.bugs.
}
