% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcra.R
\name{print.lcra_pi}
\alias{print.lcra_pi}
\title{Printing an lcra_pi object}
\usage{
\method{print}{lcra_pi}(x, ...)
}
\arguments{
\item{x}{an object of class 'lcra_pi', see lcra() for details}

\item{...}{further arguments to print}
}
\description{
Print the lcra pi output.
}
