% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{Procrustes}
\alias{Procrustes}
\title{Procrustes Projection}
\usage{
Procrustes(A, B)
}
\arguments{
\item{A}{The objective matrix to be approximated}

\item{B}{The matrix to be used to approximate}
}
\description{
This function allows you to compute the procrustes projections, min_U ||A - UB||^2 U^T U = I. It is an inner function, which shouldn't be used by users alone.
}
\keyword{procrustes}
