% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate-component.R
\name{rotate_component}
\alias{rotate_component}
\title{Rotate components}
\usage{
rotate_component(comp, angle = 0)
}
\arguments{
\item{comp}{The simulated component}

\item{angle}{The angle to be rotated of the comp}
}
\description{
This function allows you to rotate the components for every simulated component. It is an inner function that does not need to be used by user manually.
}
\keyword{rotation}
