% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-generation.R
\name{score_generation}
\alias{score_generation}
\title{Score generation}
\usage{
score_generation(dim_score, num_score, score_variance)
}
\arguments{
\item{dim_score}{The dimension of the scores.}

\item{num_score}{How many scores needed.}

\item{score_variance}{The variance of each scores.}
}
\description{
This function allows you to random generate scores to be used. It is a inner function that users do not need to use manually.
}
\keyword{score}
