% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data-generation.R
\name{simulated_data_generation}
\alias{simulated_data_generation}
\title{generate simulated data}
\usage{
simulated_data_generation(configuration_setting)
}
\arguments{
\item{configuration_setting}{The output of the configuration_setting_generation function}
}
\description{
This function allows you to generate simulated data
}
\examples{
configuration_setting = configuration_setting_generation()
data_list = simulated_data_generation(configuration_setting)
}
\keyword{simulation}
