% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-result.R
\name{visualize_result}
\alias{visualize_result}
\title{Visualization function}
\usage{
visualize_result(
  dataset,
  group,
  factor_num,
  linked_component_list,
  score_list,
  dataset_exp
)
}
\arguments{
\item{dataset}{A list containing all the datasets.}

\item{group}{A list containing the linked groups of datasets.}

\item{factor_num}{A vector containing the number of factors each group should have.}

\item{linked_component_list}{A list containing the estimated linked component.}

\item{score_list}{A list containing the list of scores.}

\item{dataset_exp}{A list containing the explanation arguments that can be used to visualize the result}
}
\description{
This function allows you to visualize the result of LCA
}
\keyword{visualization}
